requirements.txt
streamlit
google-generativeai
fpdf2
matplotlib
